renamed PersistentProgress
