+++ Assets/_Project/Scripts/Services/PersistentProgress/IProgressUpdater.cs
@@ -1,4 +1,4 @@
-﻿namespace _Project.PlayerProgress
+﻿namespace _Project.PersistentProgress
 {
     public interface IProgressUpdater : ISavedProgressReader
     {
