+++ Assets/_Project/Scripts/Services/StaticData/_Project.StaticData.asmdef
@@ -1,3 +1,16 @@
 {
-	"name": "_Project.StaticData"
+    "name": "_Project.StaticData",
+    "rootNamespace": "_Project.StaticData",
+    "references": [
+        "GUID:4d4206dc6ce48c748bd84f3aa1d7174f"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
 }
